notifier: fix reboot scheduling at ends of months
Previously we were just adding n days to the date string which didn't work
at the end of the month. This caused the notification about the reboot
to not contain the date of the scheduled reboot. Calculating with unix
time is much more reliable.

This solution was proposed by Andrew Latham in https://gitlab.nic.cz/turris/user-notify/-/issues/10#note_261465